+++ src/main/java/com/munecting/api/domain/address/controller/AddressController.java
@@ -1,19 +1,13 @@
 package com.munecting.api.domain.address.controller;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.munecting.api.domain.comment.dto.CommentRequestDto;
 import com.munecting.api.domain.address.service.AddressService;
 import com.munecting.api.global.common.dto.response.ApiResponse;
-import com.munecting.api.global.common.dto.response.Status;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.json.JSONObject;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
