+++ src/main/java/com/munecting/api/domain/spotify/dto/response/ArtistResponseDto.java
@@ -1,12 +1,9 @@
-package com.munecting.api.domain.spotify.dto;
+package com.munecting.api.domain.spotify.dto.response;
 
 import com.wrapper.spotify.model_objects.specification.Artist;
 import com.wrapper.spotify.model_objects.specification.ArtistSimplified;
 import com.wrapper.spotify.model_objects.specification.Image;
-import java.util.List;
 import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
 
 @Builder
 public record ArtistResponseDto(
